🔨 refactor: 일반 흘러온 편지 디렉토리를 NormalReception 세부 디렉토리로 분리
