+++ src/pages/LetterReceptionPage/NormalReception/ReceivedLetter/index.tsx
