+++ src/pages/LetterReceptionPage/NormalReception/ReceivedLetter/styles.ts
