+++ src/pages/LetterReceptionPage/NormalReception/components/LetterContent.tsx
