+++ src/pages/LetterReceptionPage/index.tsx
@@ -1,26 +1,28 @@
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { Suspense } from 'react';
+import { useSuspenseQuery } from '@tanstack/react-query';
 import LoadingSpinner from '@/components/LoadingSpinner';
-import createFunnel from '@/components/Funnel/createFunnel';
-import { ROUTER_PATHS } from '@/router';
 import UnknownErrorBoundary from '@/components/ErrorBoundary/UnknownErrorBoundary';
 import LetterAccessFallback from '@/components/ErrorBoundary/fallback/LetterAccessFallback';
-import ReceivedLetter from './ReceivedLetter';
-import ReplyToLetter from './ReplyToLetter';
-import ReceptionHeader from './components/ReceptionHeader';
+import letterOptions from '@/api/letter/queryOptions';
 import style from './styles';
+import NormalReception from './NormalReception';
 
-const { Funnel, Step, useFunnel } = createFunnel([
-  'ReceivedLetter',
-  'ReplyToLetter',
-] as const);
-
-const LetterReceptionPage = () => {
-  const navigate = useNavigate();
-  const { step, setStep, toPrev } = useFunnel();
-
+const SuspensedPage = () => {
   const { letterId } = useParams();
 
+  const { data: letter } = useSuspenseQuery(
+    letterOptions.singleReception(Number(letterId)),
+  );
+
+  if (letter.letterType === 'Onboarding') {
+    return <div>온보딩 고고</div>;
+  } else {
+    return <NormalReception />;
+  }
+};
+
+const LetterReceptionPage = () => {
   return (
     <UnknownErrorBoundary FallbackComponent={LetterAccessFallback}>
       <Suspense
@@ -31,30 +33,7 @@ const LetterReceptionPage = () => {
           </div>
         }
       >
-        <div css={style.container}>
-          <ReceptionHeader
-            onClickPrev={
-              step === 'ReceivedLetter'
-                ? () => navigate(ROUTER_PATHS.ROOT)
-                : toPrev
-            }
-            letterId={Number(letterId)}
-          />
-          <Funnel step={step}>
-            <Step name="ReceivedLetter">
-              <ReceivedLetter
-                letterId={Number(letterId)}
-                onNext={() => setStep('ReplyToLetter')}
-              />
-            </Step>
-            <Step name="ReplyToLetter">
-              <ReplyToLetter
-                letterId={Number(letterId)}
-                onPrev={() => setStep('ReceivedLetter')}
-              />
-            </Step>
-          </Funnel>
-        </div>
+        <SuspensedPage />
       </Suspense>
     </UnknownErrorBoundary>
   );
