+++ src/pages/LetterReceptionPage/styles.ts
@@ -1,12 +1,6 @@
 import { css } from '@emotion/react';
 
 const style = {
-  container: css`
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-  `,
   loadingSpinner: css`
     display: flex;
     flex-direction: column;
